# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dwizzyBRAIN API",
    "version": "1.8.0",
    "description": "Standardized read API for market snapshots, OHLCV candles, aggregated tickers, arbitrage signals, quant pattern search, DeFi registry reads, news feeds, and auth/session endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/v1/news": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "List news articles",
        "operationId": "listNewsArticles",
        "parameters": [
          {
            "$ref": "#/components/parameters/NewsLimit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/NewsCategory"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated news article list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsArticleListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/news/{value}": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get a news article or category feed",
        "operationId": "getNewsValue",
        "parameters": [
          {
            "$ref": "#/components/parameters/NewsValue"
          },
          {
            "$ref": "#/components/parameters/NewsLimit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "News article detail or category feed",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/NewsArticleResponse"
                    },
                    {
                      "$ref": "#/components/schemas/NewsArticleListResponse"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/news/coin/{coin_id}": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get news articles for a coin",
        "operationId": "getNewsByCoin",
        "parameters": [
          {
            "$ref": "#/components/parameters/CoinID"
          },
          {
            "$ref": "#/components/parameters/NewsLimit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "News articles associated with a coin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsArticleListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/news/trending": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get trending news analytics",
        "operationId": "getTrendingNews",
        "parameters": [
          {
            "$ref": "#/components/parameters/NewsWindow"
          },
          {
            "$ref": "#/components/parameters/TrendingLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Trending news response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsTrendingResponseEnvelope"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/quant/pattern": {
      "get": {
        "tags": [
          "Quant"
        ],
        "summary": "Search similar historical candles by fingerprint",
        "operationId": "getQuantPatternMatches",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "description": "Exchange symbol such as BTC/USDT.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": true,
            "description": "Candle timeframe such as 1m, 1h, or 4h.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of pattern matches to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "min_matches",
            "in": "query",
            "required": false,
            "description": "Confidence gate threshold for the similarity sample size.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pattern search response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "low_confidence",
                        "query",
                        "matches",
                        "outcomes"
                      ],
                      "properties": {
                        "low_confidence": {
                          "type": "boolean"
                        },
                        "query": {
                          "type": "object"
                        },
                        "matches": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "outcomes": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/quant/signals/latest": {
      "get": {
        "tags": [
          "Quant"
        ],
        "summary": "Get the latest quant signal for a market",
        "operationId": "getQuantLatestSignal",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "description": "Exchange symbol such as BTC/USDT.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": true,
            "description": "Candle timeframe such as 1m, 1h, or 4h.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "description": "Optional exchange filter such as binance.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest signal response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuantSignalLatestResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/quant/signals": {
      "get": {
        "tags": [
          "Quant"
        ],
        "summary": "List recent quant signals for a market",
        "operationId": "getQuantSignalHistory",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "description": "Exchange symbol such as BTC/USDT.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": true,
            "description": "Candle timeframe such as 1m, 1h, or 4h.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "description": "Optional exchange filter such as binance.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of signals to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal history response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuantSignalHistoryResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/quant/signals/summary": {
      "get": {
        "tags": [
          "Quant"
        ],
        "summary": "Summarize recent quant signals for a market",
        "operationId": "getQuantSignalSummary",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "description": "Exchange symbol such as BTC/USDT.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": true,
            "description": "Candle timeframe such as 1m, 1h, or 4h.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "description": "Optional exchange filter such as binance.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Sample size used for the summary.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal summary response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/QuantSignalSummaryResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "CoinID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Canonical coin identifier such as bitcoin or ethereum.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "description": "Zero-based offset into the collection.",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
...
      "NewsLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of news articles to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "NewsCategory": {
        "name": "category",
        "in": "query",
        "required": false,
        "description": "Optional article category filter such as defi, regulation, or macro.",
        "schema": {
          "type": "string"
        }
      },
      "NewsValue": {
        "name": "value",
        "in": "path",
        "required": true,
        "description": "Either a numeric article id or a category slug.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "NewsWindow": {
        "name": "window",
        "in": "query",
        "required": false,
        "description": "Lookback window for trending analytics.",
        "schema": {
          "type": "string",
          "default": "24h"
        }
      },
      "TrendingLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of trending items to return per section.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50,
          "default": 10
        }
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "NewsArticleMetadata": {
        "type": "object",
        "properties": {
          "summary_short": {
            "type": "string"
          },
          "summary_long": {
            "type": "string"
          },
          "key_points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sentiment": {
            "type": "string"
          },
          "sentiment_score": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "importance_score": {
            "type": "number"
          },
          "is_breaking": {
            "type": "boolean"
          },
          "breaking_type": {
            "type": "string"
          },
          "model_used": {
            "type": "string"
          },
          "processing_latency_ms": {
            "type": "integer"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewsArticleSummary": {
        "type": "object",
        "required": [
          "id",
          "external_id",
          "source",
          "title",
          "published_at",
          "fetched_at",
          "is_processed"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "external_id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "source_name": {
            "type": "string"
          },
          "source_credibility": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "body_preview": {
            "type": "string"
          },
          "full_url": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time"
          },
          "is_processed": {
            "type": "boolean"
          },
          "metadata": {
            "$ref": "#/components/schemas/NewsArticleMetadata"
          }
        }
      },
      "NewsArticleDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NewsArticleSummary"
          },
          {
            "type": "object",
            "properties": {
              "entities": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewsEntity"
                }
              },
              "price_impact": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewsPriceImpact"
                }
              }
            }
          }
        ]
      },
      "NewsArticleListMeta": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "category": {
            "type": "string"
          },
          "coin_id": {
            "type": "string"
          }
        }
      },
      "NewsArticleListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsArticleSummary"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/NewsArticleListMeta"
          }
        }
      },
      "NewsArticleResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NewsArticleDetail"
          }
        }
      },
...
      "NewsTrendingArticle": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NewsArticleSummary"
          },
          {
            "type": "object",
            "properties": {
              "importance_score": {
                "type": "number"
              }
            }
          }
        ]
      },
      "NewsTrendingResponse": {
        "type": "object",
        "required": [
          "window",
          "generated_at",
          "coins",
          "protocols",
          "categories",
          "top_articles"
        ],
        "properties": {
          "window": {
            "type": "string"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          },
          "coins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsTrendingCoin"
            }
          },
          "protocols": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsTrendingProtocol"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsTrendingCategory"
            }
          },
          "top_articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsTrendingArticle"
            }
          }
        }
      },
      "NewsTrendingResponseEnvelope": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NewsTrendingResponse"
          }
        }
      },
      "QuantSignalQuery": {
        "type": "object",
        "required": [
          "symbol",
          "timeframe"
        ],
        "properties": {
          "symbol": {
            "type": "string"
          },
          "timeframe": {
            "type": "string"
          },
          "exchange": {
            "type": "string"
          }
        }
      },
      "QuantSignalRecord": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "coin_id",
          "exchange",
          "symbol",
          "timeframe",
          "signal_type",
          "volume_spike",
          "price_deviation"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "coin_id": {
            "type": "string"
          },
          "exchange": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "timeframe": {
            "type": "string"
          },
          "quant_score": {
            "type": "number"
          },
          "signal_type": {
            "type": "string"
          },
          "strength": {
            "type": "string"
          },
          "rsi_14": {
            "type": "number"
          },
          "macd_line": {
            "type": "number"
          },
          "macd_signal": {
            "type": "number"
          },
          "macd_hist": {
            "type": "number"
          },
          "bb_upper": {
            "type": "number"
          },
          "bb_lower": {
            "type": "number"
          },
          "bb_mid": {
            "type": "number"
          },
          "ema_9": {
            "type": "number"
          },
          "ema_21": {
            "type": "number"
          },
          "ema_200": {
            "type": "number"
          },
          "funding_rate": {
            "type": "number"
          },
          "funding_sentiment": {
            "type": "string"
          },
          "volume_spike": {
            "type": "boolean"
          },
          "price_deviation": {
            "type": "boolean"
          },
          "anomaly_score": {
            "type": "number"
          },
          "price_at_signal": {
            "type": "number"
          }
        }
      },
      "QuantSignalLatestResponse": {
        "type": "object",
        "required": [
          "query",
          "signal"
        ],
        "properties": {
          "query": {
            "$ref": "#/components/schemas/QuantSignalQuery"
          },
          "signal": {
            "$ref": "#/components/schemas/QuantSignalRecord"
          }
        }
      },
      "QuantSignalHistoryResponse": {
        "type": "object",
        "required": [
          "query",
          "limit",
          "items"
        ],
        "properties": {
          "query": {
            "$ref": "#/components/schemas/QuantSignalQuery"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuantSignalRecord"
            }
          }
        }
      },
      "QuantSignalSummaryResponse": {
        "type": "object",
        "required": [
          "query",
          "limit",
          "count",
          "signal_type_counts"
        ],
        "properties": {
          "query": {
            "$ref": "#/components/schemas/QuantSignalQuery"
          },
          "limit": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          },
          "latest": {
            "$ref": "#/components/schemas/QuantSignalRecord"
          },
          "avg_quant_score": {
            "type": "number"
          },
          "avg_funding_rate": {
            "type": "number"
          },
          "avg_anomaly_score": {
            "type": "number"
          },
          "volume_spike_rate": {
            "type": "number"
          },
          "price_deviation_rate": {
            "type": "number"
          },
          "signal_type_counts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "strength_counts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        }
      }
    }
  }
}